changed ingredient quantity to int, rather than decimal

+++ db/tables.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS steps;
 DROP TABLE IF EXISTS quantities;
 DROP TABLE IF EXISTS measurements;
 DROP TABLE IF EXISTS ingredients;
-DROP TABLE IF EXISTS recipes; 
+DROP TABLE IF EXISTS recipes;
 
 CREATE TABLE recipes(
   id SERIAL8 PRIMARY KEY,
@@ -25,7 +25,7 @@ CREATE TABLE quantities(
   recipeId INT8 REFERENCES recipes(id) ON DELETE CASCADE,
   ingredientId INT8 REFERENCES ingredients(id) ON DELETE CASCADE,
   measurementId INT8 REFERENCES measurements(id) ON DELETE CASCADE,
-  ingredientQuantity DECIMAL(8,2)
+  ingredientQuantity INT
 );
 
 CREATE TABLE steps(
